refactor: revert renaming of lib-franklin.js
