+++ 404.html
@@ -11,7 +11,7 @@
   <meta property="og:title" content="Page not found">
   <script src="/solutions/scripts/scripts.js" type="module" crossorigin="use-credentials"></script>
   <script type="module">
-    import { decorateButtons, sampleRUM } from '/solutions/scripts/aem.js';
+    import { decorateButtons, sampleRUM } from '/solutions/scripts/lib-franklin.js';
 
     window.addEventListener('load', async () => {
       const response = await fetch('/solutions/404.plain.html');
