+++ head.html
@@ -1,6 +1,6 @@
 <meta name="referrer" content="no-referrer-when-downgrade" />
 <meta name="viewport" content="width=device-width, initial-scale=1"/>
-<script src="/solutions/scripts/aem.js" type="module"></script>
+<script src="/solutions/scripts/lib-franklin.js" type="module"></script>
 <script src="/solutions/scripts/utils.js" type="module"></script>
 <script src="/solutions/scripts/scripts.js" type="module"></script>
 <link rel="stylesheet" href="/solutions/styles/styles.css"/>
