+++ solutions/blocks/quote-carousel/quote-carousel.js
@@ -1,5 +1,5 @@
 import { createTag } from '../../scripts/utils.js';
-import { decorateIcons } from '../../scripts/aem.js';
+import { decorateIcons } from '../../scripts/lib-franklin.js';
 
 const SLIDE_PREFIX = 'carousel-slide-';
 const CONTROL_PREFIX = 'carousel-control-';
