+++ solutions/blocks/sticky-navigation/sticky-navigation.js
@@ -1,4 +1,4 @@
-import { STICKY_NAVIGATION_DATASET_KEY } from '../../scripts/aem.js';
+import { STICKY_NAVIGATION_DATASET_KEY } from '../../scripts/lib-franklin.js';
 
 function scrollToAnchorWithOffset(anchorId) {
   const anchorElement = document.getElementById(anchorId);
