+++ solutions/scripts/lib-franklin.js
